fix count of days in countdowns (#1764)
* fix count of days in countdowns

* single version of calculateCountdown
