+++ src/components/Shared/Countdown.tsx
@@ -2,6 +2,7 @@ import * as moment from "moment";
 import * as React from "react";
 
 import * as css from "./Countdown.scss";
+import { calculateCountdown } from "lib/util";
 
 interface IProps {
   toDate: Date | moment.Moment;
@@ -10,7 +11,6 @@ interface IProps {
 }
 
 interface IState {
-  years: number;
   days: number;
   hours: number;
   min: number;
@@ -23,13 +23,7 @@ export default class Countdown extends React.Component<IProps, IState> {
   constructor(props: IProps) {
     super(props);
 
-    this.state = this.calculateCountdown(this.props.toDate) || {
-      years: 0,
-      days: 0,
-      hours: 0,
-      min: 0,
-      seconds: 0,
-    };
+    this.state = calculateCountdown(this.props.toDate);
   }
 
   public componentDidMount() {
@@ -41,9 +35,9 @@ export default class Countdown extends React.Component<IProps, IState> {
 
       // update every five seconds
       this.interval = setInterval(() => {
-        const date = this.calculateCountdown(this.props.toDate);
-        if (date) {
-          this.setState(date);
+        const countdown = calculateCountdown(this.props.toDate);
+        if (!countdown.complete) {
+          this.setState(countdown);
         } else {
           this.stop();
           if (this.props.onEnd) {
@@ -56,28 +50,6 @@ export default class Countdown extends React.Component<IProps, IState> {
 
   public componentWillUnmount() {
     this.stop();
-  }
-
-  public calculateCountdown(endDate: Date | moment.Moment): IState {
-    const endDateMoment = moment(endDate); const now = new Date();
-
-    const diff = endDateMoment.diff(now);
-
-    // clear countdown when date is reached
-    if (diff <= 0) {
-      return null;
-    }
-
-    const duration = moment.duration(diff);
-    const timeLeft = {
-      years: duration.years(),
-      days: duration.days(),
-      hours: duration.hours(),
-      min: duration.minutes(),
-      seconds: duration.seconds(),
-    };
-
-    return timeLeft;
   }
 
   public stop() {
